% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{calculate_lip_usage}
\alias{calculate_lip_usage}
\title{Calculate LiP-MS Usage (Corrected logFC)}
\usage{
calculate_lip_usage(
  qf,
  i_lip,
  i_tc,
  contrasts,
  fcol_lip = "Protein.Group",
  fcol_tc = "Protein.Group",
  satterthwaite = FALSE
)
}
\arguments{
\item{qf}{A \code{QFeatures} object.}

\item{i_lip}{\code{character(1)}: The name of the assay containing LiP-level models.}

\item{i_tc}{\code{character(1)}: The name of the assay containing TC-level models.}

\item{contrasts}{\code{character()}: A vector of contrast names (e.g., "A - B") 
that exist in the \code{rowData} of both assays.}

\item{fcol_lip}{\code{character(1)}: Column in \code{rowData(qf[[i_lip]])} 
used for matching to proteins. Default is "Protein.Group".}

\item{fcol_tc}{\code{character(1)}: Column in \code{rowData(qf[[i_tc]])} 
identifying proteins. Default is "Protein.Group".}

\item{satterthwaite}{\code{logical(1)}: Whether to use the Satterthwaite 
approximation for degrees of freedom. Default is \code{FALSE}.}
}
\value{
A \code{QFeatures} object with updated \code{rowData} in the \code{i_lip} assay.
}
\description{
This function corrects LiP peptide log-fold changes (logFC) by subtracting the 
total protein (TC) logFC. It performs error propagation for standard errors 
and re-calculates p-values for the "usage" (the peptide change relative to 
the protein change).
}
