% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{create_qfeat_}
\alias{create_qfeat_}
\title{create_qfeat_}
\usage{
create_qfeat_(annotation_df, report_file1, report_file2 = NULL, diann_flag)
}
\arguments{
\item{annotation_df}{A data frame containing sample metadata (colData).}

\item{report_file1}{A data frame containing quantitative proteomics data.}

\item{report_file2}{Optional; secondary report file (default is NULL).}

\item{diann_flag}{Logical; TRUE if data is from DIA-NN, FALSE for Spectronaut.}
}
\value{
A list with status, error message, and the resulting QFeatures object.
}
\description{
Converts raw data frames into a QFeatures object, calculates group-specific 
detection heuristics, and performs initial quality filtering.
}
