% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{make_mds}
\alias{make_mds}
\title{Create MDS Plot Grob}
\usage{
make_mds(pe)
}
\arguments{
\item{pe}{A \code{QFeatures} object containing \code{precursors_lip_norm} 
and \code{precursors_tc_norm} assays.}
}
\value{
A \code{gtable} (grob) object representing the MDS plot.
}
\description{
Performs Multi-Dimensional Scaling (MDS) on joined assays within a QFeatures 
object, calculates variance explained, and returns a rendered ggplot grob.
}
