% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{msqrob_model}
\alias{msqrob_model}
\title{Fit msqrob2 Models and Test Contrasts}
\usage{
msqrob_model(pe, params, layer)
}
\arguments{
\item{pe}{A \code{QFeatures} object.}

\item{params}{A list containing \code{formula} (string) and \code{comparisons} (character vector).}

\item{layer}{Character. The name of the assay in \code{pe} to model.}
}
\value{
A list with the following components:
\itemize{
  \item \code{status}: Integer (0 for success, 1 for error).
  \item \code{error}: Character string containing error messages.
  \item \code{q_feat}: The updated \code{QFeatures} object with models and test results.
  \item \code{contr_exp}: The contrast matrix used for testing.
}
}
\description{
Fits robust linear models to proteomics data using the msqrob2 framework. 
It performs hypothesis testing based on a user-defined formula and 
contrast matrix, handling peptide-level or protein-level quantification.
}
\author{
Andrea Argentini
}
