% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_function.R
\name{validate_folder}
\alias{validate_folder}
\title{Validate report folder path}
\usage{
validate_folder(report_folder)
}
\arguments{
\item{report_folder}{Character scalar. Path to the folder to validate/create.}
}
\value{
Logical TRUE if the folder is valid/created (invisibly); otherwise the function stops with an error.
}
\description{
Validate and prepare the report folder. Creates the folder (and a "Result" subfolder) if it does not exist. Returns TRUE when the folder is available; otherwise the function will stop with an error.
}
\author{
andrea Argentini
}
