% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{calculate_coverages}
\alias{calculate_coverages}
\title{Calculate Sequence Coverage per Protein}
\usage{
calculate_coverages(report)
}
\arguments{
\item{report}{A \code{data.frame} or \code{tibble} containing "Accession", 
"length", "start", and "end".}
}
\value{
A list with the following components:
\itemize{
  \item \code{status}: Integer (0 for success, 1 for error).
  \item \code{error}: Character string containing error messages.
  \item \code{result}: The original data frame with an added "coverage" column.
}
}
\description{
Calculates the fraction of each protein sequence covered by the detected 
peptides. Overlapping peptide regions are reduced to unique amino acid 
positions to ensure accurate coverage mapping.
}
\author{
Andrea Argentini
}
