% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{check_design_requirement}
\alias{check_design_requirement}
\title{check_design_requirement}
\usage{
check_design_requirement(df, required_cols)
}
\arguments{
\item{df}{Input data frame where features need to be checked}

\item{required_cols}{Character vector of required column names}
}
\value{
A list with elements:
  \item{status}{integer; 0 if no error, 1 if an error was found}
  \item{error}{character; error message when status is 1, otherwise an empty string}
}
\description{
This function performs the following checks on the design file:
 1) Check required columns exist
 2) Check Run column has no file extensions
 3) Check Pipeline column values
}
\author{
Andrea Argentini
}
