% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{compute_usage}
\alias{compute_usage}
\title{compute_usage}
\usage{
compute_usage(q_feat)
}
\arguments{
\item{q_feat}{A \code{QFeatures} object.}
}
\value{
A list with the following components:
\itemize{
  \item \code{error}: Character string containing error messages, if any.
  \item \code{status}: Integer (0 for success, 1 for error).
  \item \code{result}: The updated \code{QFeatures} object containing the "precursors_lip_usage" assay.
}
}
\description{
This function calculates the relative peptide usage by subtracting the 
corresponding protein-level abundance (Total Control) from normalized 
peptide intensities. This step adjusts for changes in total protein expression 
to isolate changes in protein conformation or accessibility.
}
\details{
Compute LiP-MS Usage
}
