% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{normalization_scaling_factor}
\alias{normalization_scaling_factor}
\title{Log-Transformation, Median Normalization, and Protein Aggregation}
\usage{
normalization_scaling_factor(q_feat)
}
\arguments{
\item{q_feat}{A \code{QFeatures} object containing "precursors_lip" and "precursors_tc" assays.}
}
\value{
A list with the following components:
\itemize{
  \item \code{error}: Character string containing error messages, if any.
  \item \code{status}: Integer (0 for success, 1 for error).
  \item \code{result}: The updated \code{QFeatures} object.
}
}
\description{
Performs log2 transformation on LiP and TC assays, calculates sample-based 
normalization factors using common features, and aggregates TC precursors 
to the protein level.
}
