% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{pep_char}
\alias{pep_char}
\title{Characterize Peptide Properties and Mapping}
\usage{
pep_char(table, prot_seq = "Protein.Sequence", pep_seq = "Stripped.Sequence")
}
\arguments{
\item{table}{A data frame or tibble.}

\item{prot_seq}{Character. Column name for protein sequence (default "Protein.Sequence").}

\item{pep_seq}{Character. Column name for stripped peptide sequence (default "Stripped.Sequence").}
}
\value{
A data frame with additional columns for sequence properties.
}
\description{
Annotates a data frame of peptides with biochemical and positional properties 
such as missed cleavages, mapping positions, and trypticity.
}
